Edition02/Hardcover/Hardcover_6x9_ed2_v1_01.odt:add
This is an attempt to make the book with LO. added revisions:

* The Search for the Enchanted Balls
* The Camel in the Desert
* The Ant and the Cliff

LO is a little easier to manage but still a bit too unpredictable for my
tastes

Also included with this commit is an LO "Master Document" and
"subdocuments".

https://help.libreoffice.org/6.2/en-US/text/swriter/guide/globaldoc_howtos.html?&DbPAR=WRITER&System=UNIX

It was interesting but I think it will complicate things too much. And
in less than 30 minutes of trying it out, had a LO crash.

[skip ci]